+++ Main App/ViewControllers/TimerRuntimeViewController.swift
@@ -540,11 +540,12 @@ class TimerRuntimeViewController: A_TimerNavBaseController {
     /* ################################################################## */
     /**
      */
-    @IBAction func stopButtonHit(_ sender: Any) {
+    @IBAction func stopButtonHit(_ sender: Any! = nil) {
         if .Alarm == self.timerObject.timerStatus {
             self.resetTimer()
             if 0 <= Timer_AppDelegate.appDelegateObject.appState.nextTimer, let myTabController = Timer_AppDelegate.appDelegateObject.mainTabController {
                 Timer_AppDelegate.appDelegateObject.timerEngine.stopTimer()
+                myTabController.timerEngine.autoStartNextSelectedTimer = true   // We want the next selected timer to start immediately.
                 myTabController.timerEngine.selectedTimerIndex = Timer_AppDelegate.appDelegateObject.appState.nextTimer
             }
         }
@@ -591,11 +592,7 @@ class TimerRuntimeViewController: A_TimerNavBaseController {
      */
     @IBAction func tapInView(_ sender: Any) {
         if .Alarm == self.timerObject.timerStatus {
-            self.resetTimer()
-            if 0 <= Timer_AppDelegate.appDelegateObject.appState.nextTimer, let myTabController = Timer_AppDelegate.appDelegateObject.mainTabController {
-                Timer_AppDelegate.appDelegateObject.timerEngine.stopTimer()
-                myTabController.timerEngine.selectedTimerIndex = Timer_AppDelegate.appDelegateObject.appState.nextTimer
-            }
+            self.stopButtonHit()
         } else {
             self.pauseButtonHit()
         }