+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Prepared for open-source (Equivalent of cleaning up before the in-laws visit).
 - Improved the Spanish localization slightly, by shortening the string for the "Show Controls" button.
 - Since the iPod touch also does not support vibrate, I switched the vibrate detection to iPhone-only.
+- I now immediately start the cascaded timer when it is selected.
 
 ***Version 1.0.0.3000* ** *(January 21, 2019)*
 - Some basic refactoring to ensure that the main thread is used, and that we use optional chaining for delegate calls.